requirements.txt
alembic==1.4.2
amqp==2.5.2
apache-airflow==1.10.10
apispec==1.3.3
argcomplete==1.11.1
attrs==19.3.0
autopep8==1.5.2
Babel==2.8.0
billiard==3.6.3.0
boto3==1.13.6
botocore==1.16.6
cached-property==1.5.1
cattrs==0.9.2
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colorlog==4.0.2
configparse==0.1.5
configparser==3.5.3
croniter==0.3.31
defusedxml==0.6.0
dill==0.3.1.1
dnspython==1.16.0
docutils==0.15.2
email-validator==1.1.0
Faker==4.0.3
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.1
flask-swagger==0.2.13
Flask-WTF==0.14.3
flower==0.9.4
funcsigs==1.0.2
future==0.18.2
graphviz==0.14
gunicorn==19.10.0
humanize==0.5.1
idna==2.9
importlib-metadata==1.6.0
influxdb==5.3.0
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.5
json-merge-patch==0.2
jsonschema==3.2.0
kombu==4.6.8
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.1.2
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.0
msgpack==0.6.1
mysql-connector-python==8.0.21
numpy==1.18.4
pandas==0.25.3
pathtools==0.1.2
pendulum==1.4.4
prison==0.1.3
protobuf==3.13.0
psutil==5.7.0
psycopg2-binary==2.8.5
pycodestyle==2.5.0
Pygments==2.6.1
PyJWT==1.7.1
PyMySQL==0.9.3
pyrsistent==0.16.0
python-daemon==2.1.2
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python3-openid==3.1.0
pytz==2020.1
pytzdata==2019.3
PyYAML==5.3.1
requests==2.23.0
s3transfer==0.3.3
setproctitle==1.1.10
six==1.14.0
SQLAlchemy==1.3.16
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.5
tabulate==0.8.7
tenacity==4.12.0
termcolor==1.1.0
text-unidecode==1.2
thrift==0.13.0
tornado==5.1.1
typing-extensions==3.7.4.2
tzlocal==1.5.1
ua-parser==0.10.0
unicodecsv==0.14.1
urllib3==1.25.9
user-agents==2.1
vine==1.3.0
watchdog==0.10.2
Werkzeug==0.16.1
WTForms==2.3.1
yapf==0.30.0
zipp==3.1.0
zope.deprecation==4.4.0
